requirements.txt
GitPython
docker
pyyaml